refactor: use decimal instead of float

+++ Runtime/Communication/Lobby/Models/ToLobby/RequestRollings.cs
@@ -21,6 +21,6 @@ namespace Communication.Lobby.Models.ToLobby
         [property: Key(0)] Guid CoinId,
         [property: Key(1)] string Prize,
         [property: Key(2)] uint PlayersCount,
-        [property: Key(3)] float[] PrizeDistribution);
+        [property: Key(3)] decimal[] PrizeDistribution);
 
 }
